[haproxy-ova] Data Plane API as SystemD service
This patch changes the way the Data Plane API is executed. Previously
it was hosted as an HAProxy "program," but that caused trouble due to
the resource consumption and eventual crashing of the Data Plane API
process. This change should help manage the resources allocated to the
Data Plane API process and ensure its lifecycle does not impact that of
the HAProxy process by limited CPU shares allocated to the Data Plane
API process.