Fix a typo in systemd restart directive for dataplaneapi
It was observed in a live setup that when dataplaneapi crashes or
goes OOM, it was not automatically restarted by systemd. This was
because the systemd restart policy is case sensitive causing systemd
to ignore the option that is set in the dataplaneapi service file.
This change corrects the issue.