Use link files for early ethernet device initialization
udevd is the device responsible for early device initialization,
which, among other things, governs device naming and idenitfication.

During reboot it was observed that udevd is in the middle of device
updates while systemd-networkd is booting. This causes networkd to
think that it a device is temporarily unmanaged, causing settings
to be reset to default in some cases.

This change anchors both daemons to a single source of truth so that
udevd is aware of these devices and should not attempt to configure
them further. Networkd should now always recognize these devices as
those it should manage.