Fix route table configuration scripts
A previous change performed some refactoring in this area to
update the ethernet names and remove dependence on PCI bus slot as
a key used throughout the configuration scripts. This shortened
the number of parameters required in some functions. This was
unaccounted for in the change resulting in broken GC deployments
because the Supervisor was unable to route to the guest clusters IP.

This change simply corrects the issues in these functions.