Ensure network link files participate in initramfs
After the move to network link files, we still observed that udevd
was attempting to rename the ethernet devices. The devices would
eventually end up with the correct name, but such a rename is not
required and can confuse systemd-networkd into thinking it should
not manage devices.

After investigation it was determined that in order to get udevd
to honor network link names in link files, we need to ensure
that these files are part of early user world initialization as
part of ramfs. These files only get included if they are included
inside /usr/lib/systemd.