Move firstboot flag after environmental checks
If we encounter a situation where we do not have the required
prerequisites of ovf environment and metadata, we prevent future
firstboots from occurring, which would cause a permanent
misconfiguration. This change ensures we bail out early before
setting the configurations so that firstboot always runs only when
we have the required prerequisites.