Fix inverted return code
A recent change accidentally inverted a return code checking for
invalid MAC addresses. This check served as a purpose of detecting
whether we used a 2 nic or a 3 nic setup.

The result is that we caused cloud-init to configure NICs invalidly
causing network issues for the appliance. This fix reverts the
change.