Remove check for existing userdata
Userdata may exist when the ovf-to-cloud-init script runs. Because
the userdata is populated by us, it was assumed that any pre-existing
userdata is a custom configuration that would be unsupported.
However, these checks are not needed given that we overwrite any
pre-existing userdata in the very next function when we publish new
userdata.

This change removes the check so that re-running cloud-init becomes
possible without changing code.