Allow for graceful deletions in case of missing routes and route tables
If a route is removed from the system and not from the route table
file, then the route table service will fail to process new routes.
Instead, if we ignore to remove any failed routes, we can still
recover by adding new routes to the system if required.

This change logs any errors in removing route or route tables instead
of outright failure so that the system can recover by readding the
expected routes.