Ensure cert and key is created from new-cert and new-ca
It was observed that openssl can occasionally output an error
code that does not invoke the retry script if not enough system
entropy exists. This commit places a simple sanity check at the end
of the scripts returning error if a file was unable to be generated
for any reason.