DOC: Update the README to use Makefile for Docker images
The build process has been updated to use a Makefile. This patch
updates the README to reflect those changes for the Docker images.
It also updates the frontend/backends in the example to use
`mode http` and removes the `check-ssl` option from the backend
servers, which is not needed because they are running over plain HTTP.