Do not generate new certs if one already exists
Cert generation is added to cloud init which will generate
new certificates to be used by DPAPI to communicate with the LBAPI.
However, after we fixed a recent patch to actually force the function
to work correctly, this script now clobbers any existing certs when
cloud-init is run. This can be issue if cloud-init is ever re-run
or if a cert is already present before cloud-init is ran in the
first place.

This commit adds a check to avoid generating certificates if they
already exist and the appropriate flag is specified. This is safer
behavior than clobbering certificates at-will in most conditions.