Add support bundle collection script
When debugging it's great to have a unified method of collecting
necessary support bundles for the VMware environment.

This commit adds a script to collect support bundles. It collects
information that could possibly result in a botched deployment such
as general network information, route tables, anyip configuration,
and HAProxy and Dataplane API logs.